+ Implement first version of REST handling for EntertainmentConfiguration obejcts.
This is a major step on the way to supporting Hue Entertainment Mode.

The current implementation still makes many assumptions, and has missing
features, but has been tested on an LCX005 gradient strip. With proper
entertainment mode streaming support (in future commits), this is able to make
Bifrost handle entertainment groups.