* Refactored and reworked svc crate
The svc crate had a major design flaw: It was impossible both use a
ServiceManager, and also maintain control over it.

So while it was possible to create services, start them, stop them, and get the
status of them, this required that everything happened from a single task.

In other words, it was not reasonably possible to share a ServiceManager behind
a mutex, since the ServiceManager itself would need mutex access to function.

With this refactor, the ServiceManager communicates entirely over channels, and
all operations happen as request/response-style messages.