+ Implement support for upnp-related "/description.xml" endpoint.
This endpoint helps some clients find Bifrost on the network.

For example, the android app "hueDynamic" seems to require this endpoint, but
it's likely that other programs will benefit as well.

The output has been carefully reviewed to match the original hue bridge output
in syntax, order and meaning. The only differences are formatting changes
(indentation is enabled for Bifrost), and the values of the various xml fields.