* Convert svc crate to use unbounded channels
The service manager used to use bounded channels to communicate with clients and
services, but this leads to difficult edge cases, that increase in likelyhood as
the queue size shrinks.

Increasing the queue limit only makes these problems less frequent. There is no
universal, "safe" queue size.

So instead, remove the queue size limit. This also removes the need for .await,
simplifying a lot of code.

In the future, we might find a better way to solve this, that doesn't involve
unbounded growth, but for now, this will do.