+ Implement working support for "Bridge Upgrade" endpoint
By reacting believably to POST /updater, and resetting the version checker cache
in the handler, it's now possible to "update" the Bifrost bridge "firmware"
directly from the hue app!

After "installing" the firmware, the (Bifrost) bridge indeed does respond with a
new version number. Nothing else is changed, of course, but it makes for a
smooth user experience.