* Completely rework handling of DTLS packet for entertainment mode
Avoid using buffered reader, to ensure we handle a single datagram at a
time. This turns out to be much simpler than the previous approach, while also
being much more resilient to unexpected input.

When reading a frame, we now have much better logging for errors that may occur,
such as end of stream, or timeout.